+++ app/controllers/trackingController.js
@@ -1,7 +1,0 @@
-//Vai manipular os dados de um tracker ja cadastrado
-
-module.exports = {
-    routes: (data) => {
-        console.log(data)
-    }
-}